server/requirements.txt
playwright>=1.58.0,<2.0.0
python-dotenv>=1.2.2,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy>=2.0.30,<3.0.0
asyncpg>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
email-validator>=2.2.0,<3.0.0
bcrypt>=4.2.0,<5.0.0
pyjwt>=2.9.0,<3.0.0
cryptography>=43.0.0,<45.0.0
apscheduler>=3.10.4,<4.0.0
